KrzYmod v14:
Added:
--rw# and --rz# to feed Multiply-With-Carry random generator with
different base values than the default 1 and 2. The range for both is
1-65535. This has an impact on iteration progression and may generate
different file sizes when using same amount of iterations.
--cbsfile to act as --cbs but instead read block start points and types
from file pointed instead of command line,
--cbd to dump block start points to given file and exit (if You plan to
rerun zopfli on big file and want to use --cbsfile to read that file and
speed up the process),
Changed:
--cbs now reads block start positions and accepts optional parameter of
block type. First block start position is always replaced with 0 so the
data may start with =block_type, or just a comma. Example:
0=1,3fe,fff=0,f999 or =1,3fe,fff=0,f999 or ,3fe,fff,f999.
Removed:
--cbt - replaced with --cbs extended functionality.