Add support for --legacy with --dir
Now --legacy command which includes exact functionality with original
zopfli (additional last block splitting after compression which is not
available in KrzYmod algorithm) works with multiple files (--dir).

Of course downside to using --legacy is higher memory usage, because
input file is kept in memory as a whole during compression, not read in
chunks + at most 32KB of back data as KrzYmod does. Same for output
memory, it's not dumped when ZOPFLI_MAX_OUTPUT_MEMORY is exceeded.