Avoid checking ptr for NULL before free...
... because as MAN for free states, if ptr is NULL, no operation is
performed by free. Checking for NULL ptr before free should still be
done, for example, when we clean a structure with sub-structures that
may not be always initialized or when we want to clean structure members
of a structure that can be NULL, for example, by being optional in a
function.

* now we can debug if there are any further problems with dynamic memory
usage.

- Update Readme