Malloc Hash Earlier (~4% faster build)
And Initialize it later to avoid heap-related problems (especially
visible on Odroid U3). We were causing unecessary memory usage spikes
and fragmentation because it was malloc'd and free'd every iteration
(too aggressive), where it actually needs to be malloc'd at the start of
iterations and free'd after last iteration.
This makes Zopfli a bit faster, before this commit: [4m55s], after:
[4m43s], tests done with --t1 on Windows 8.1. The speed difference gap
should be bigger on devices with slower memory access or worse memory
manager.

Reverse the order of free where possible. This is not really necessary,
but again, this is nicer for the heap.

Reduce hash initialization tries to 10 and display error message only
before exitting.