This fixed SIGSEGV on Odroid U3 for now
This should fix Odroid U3 SIGSEGV as the problem was related to disabled
unaligned access along with 64 bit being structure boundary size on
pointer structure members. Use 32 bit instead.

Remove unecessary nulling of pointer struct members, as we know that we
won't run into any double-free there anyway, so this is not needed.